add missing patches from Qemu trunk (still incomplete)
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/ar7-firmware/qemu/trunk@1171 11fd3d49-aaf6-0310-8052-ea9566248e5e

+++ cpu-all.h
@@ -667,7 +667,7 @@ static inline void stfq_be_p(void *ptr, float64 v)
 
 /* All direct uses of g2h and h2g need to go away for usermode softmmu.  */
 #define g2h(x) ((void *)((unsigned long)(x) + GUEST_BASE))
-#define h2g(x) ((target_ulong)(x - GUEST_BASE))
+#define h2g(x) ((target_ulong)((unsigned long)(x) - GUEST_BASE))
 
 #define saddr(x) g2h(x)
 #define laddr(x) g2h(x)
